Chore: Evaluate StatusView - TypeScript ++hooks (#4075)
* move statusview to your specific folder

* create showToast helper

* migrate StatusView to hooks

* fix type

* add peer deep to useEffect

* fix lint

* fix error message
