+++ app/lib/methods/helpers/showToast.ts
@@ -1,0 +1,4 @@
+import { LISTENER } from '../../../containers/Toast';
+import EventEmitter from '../../../utils/events';
+
+export const showToast = (message: string): void => EventEmitter.emit(LISTENER, { message });
